chore: rename pocketguard => fundli (#27)
