+++ README.md
@@ -1,2 +1,1 @@
-# pocketguard
-A money tracker
+# fundli
