+++ fundli/services/shared.py
