+++ fundli/tests/services_test.py
@@ -1,7 +1,7 @@
 import pytz
 from datetime import datetime
 
-from pocketguardapp.services.shared import parse_datetime
+from fundli.services.shared import parse_datetime
 
 
 def test_parse_datetime():
